requirements.txt
# FastAPI and server
fastapi>=0.109.0,<0.110.0
uvicorn[standard]>=0.27.0,<0.28.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.23,<3.0.0
psycopg2-binary>=2.9.9,<3.0.0
pgvector>=0.2.4

# Authentication and Security
passlib[bcrypt]>=1.7.4,<2.0.0
bcrypt==3.2.0  # Pin bcrypt to version compatible with passlib (has __about__ attribute)
python-jose[cryptography]>=3.3.0,<4.0.0

# Data Validation
pydantic[email]>=2.5.3,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# Environment variables
python-dotenv>=1.0.0,<2.0.0

# AI/ML Libraries (torch is installed separately in Dockerfile with CPU-only version)
transformers>=4.37.0,<5.0.0
sentence-transformers>=2.3.1,<3.0.0
numpy>=1.24.0,<2.0.0

# NLP
nltk>=3.8.1,<4.0.0

# Document Processing
PyMuPDF>=1.23.8,<2.0.0
docx2txt>=0.8,<1.0.0

# Caching
redis>=5.0.1,<6.0.0
hiredis>=2.2.3,<3.0.0